requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-WTF==1.2.1
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
# Logging enhancements
python-json-logger==2.0.7
structlog==23.2.0
# YouTube video handling
pytube==15.0.0
requests==2.32.3
# Production server
gunicorn==21.2.0
# Security
Flask-Talisman==1.1.0
# Error tracking
sentry-sdk[flask]==1.39.1
